Move ExposedDropdownMenuDefaults to common
